third_party/blink/web_tests/platform/linux/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/abs.https.any_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] abs float16 positive 0D scalar
  assert_less_than_equal: assert_array_approx_equals_ulp: test abs float16 actual 0 should be close enough to expected 49.84375 by ULP distance: expected a number less than or equal to 0 but got 21051
[FAIL] abs float16 negative 0D scalar
  assert_less_than_equal: assert_array_approx_equals_ulp: test abs float16 actual 0 should be close enough to expected 91.0625 by ULP distance: expected a number less than or equal to 0 but got 21937
[FAIL] abs float16 1D constant tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test abs float16 actual 512 should be close enough to expected 49.84375 by ULP distance: expected a number less than or equal to 0 but got 3525
[FAIL] abs float16 1D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test abs float16 actual 0 should be close enough to expected 49.84375 by ULP distance: expected a number less than or equal to 0 but got 21051
[FAIL] abs float16 2D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test abs float16 actual 0 should be close enough to expected 49.84375 by ULP distance: expected a number less than or equal to 0 but got 21051
[FAIL] abs float16 3D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test abs float16 actual 0 should be close enough to expected 49.84375 by ULP distance: expected a number less than or equal to 0 but got 21051
[FAIL] abs float16 4D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test abs float16 actual 0 should be close enough to expected 49.84375 by ULP distance: expected a number less than or equal to 0 but got 21051
[FAIL] abs int8 4D tensor
  promise_test: Unhandled rejection with value: object "TypeError: Unsupported data type, input 'input' data type int8 must be one of [float32,float16,int32]."
[FAIL] abs int64 4D tensor
  promise_test: Unhandled rejection with value: object "TypeError: Unsupported data type, input 'input' data type int64 must be one of [float32,float16,int32]."
Harness: the test ran to completion.
